Put the bus back in iDevGames

+++ deploy.sh
@@ -9,6 +9,21 @@ rust_in_docker() {
     $@
 }
 
+fast=false
+
+while getopts :f arg
+do
+  case "${arg}" in
+    f)
+      fast=true
+      ;;
+    ?)
+      echo "Invalid option -$OPTARG"
+      exit 2
+      ;;
+  esac
+done
+
 set -o xtrace
 set -x
 set -e
@@ -19,18 +34,28 @@ SERVICE_USER_NAME=mysteriousbot
 SERVICE_USER_DIR=/home/mysteriousbot
 SERVICE_NAME=mysteriousbot
 
-rust_in_docker cargo test -- --nocapture
-rust_in_docker cargo build --release
+if [[ $fast == false ]]; then
+  rust_in_docker cargo test -- --nocapture
+  rust_in_docker cargo build --release
+fi
 
 # deploy the build if successful
 if [ $? -eq 0 ]; then
-  ssh ${SERVICE_SU_USER}@${SERVICE_HOST} sudo systemctl stop ${SERVICE_NAME}
-  ssh ${SERVICE_USER_NAME}@${SERVICE_HOST} mkdir -p ${SERVICE_USER_DIR}/config
-  ssh ${SERVICE_USER_NAME}@${SERVICE_HOST} mkdir -p ${SERVICE_USER_DIR}/db
+  if [[ $fast == false ]]; then
+    ssh ${SERVICE_SU_USER}@${SERVICE_HOST} sudo systemctl stop ${SERVICE_NAME}
+    ssh ${SERVICE_USER_NAME}@${SERVICE_HOST} mkdir -p ${SERVICE_USER_DIR}/config
+    ssh ${SERVICE_USER_NAME}@${SERVICE_HOST} mkdir -p ${SERVICE_USER_DIR}/db
+  fi
+
   rsync -avzr config/mysteriousbot.yml ${SERVICE_USER_NAME}@${SERVICE_HOST}:${SERVICE_USER_DIR}/config/mysteriousbot.yml
-  rsync -avzr target/release/mysteriousbot ${SERVICE_USER_NAME}@${SERVICE_HOST}:${SERVICE_USER_DIR}/mysteriousbot
-  ssh ${SERVICE_USER_NAME}@${SERVICE_HOST} chmod +x ${SERVICE_USER_DIR}/mysteriousbot
-  ssh ${SERVICE_SU_USER}@${SERVICE_HOST} sudo systemctl start ${SERVICE_NAME}
+
+  if [[ $fast == false ]]; then
+    rsync -avzr target/release/mysteriousbot ${SERVICE_USER_NAME}@${SERVICE_HOST}:${SERVICE_USER_DIR}/mysteriousbot
+    ssh ${SERVICE_USER_NAME}@${SERVICE_HOST} chmod +x ${SERVICE_USER_DIR}/mysteriousbot
+    ssh ${SERVICE_SU_USER}@${SERVICE_HOST} sudo systemctl start ${SERVICE_NAME}
+  else
+    ssh ${SERVICE_SU_USER}@${SERVICE_HOST} sudo systemctl restart ${SERVICE_NAME}
+  fi
 else
   echo "Fix your broken build, man."
 fi
